Inject Jenkins build number into version info
#risk3 #build #ci #jenkins #warning

Jenkins build number is appended to the version name.
The Jenkins build number is combined with the minSdkVersion and other info to make the versionCode:
114100062
^---------- Version system 1
 ^^-------- minSdkVersion - Allows multi-sdk in play store
   ^------- Constant to bring code above last release (114030405)
    ^^^^^-- Jenkins build no

This commit has introduced a warning 'defaultConfig' cannot be applied to '(groovy.lang.Closure), but stackoverflow suggests this can be ignored. http://stackoverflow.com/a/31923462/270155